LPS-128505 Inline BasePreferencesImpl.
